@@ -1,12 +1,29 @@
+const path = require('path');
+
 module.exports = {
-  entry: './js/index.js',
-  mode: 'production',
+  entry: path.resolve(__dirname, 'js/index.js'),
+  mode: process.env.NODE_ENV === 'production' ? 'production' : 'development',
   output: {
-    path: __dirname + '/public/js'
+    path: path.resolve(__dirname, 'public/js'),
+    filename: 'main.js'
+  },
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.jsx']
   },
   module: {
     rules: [
-        {test: /\.js$/, use: "babel-loader"}
+      {
+        test: /\.(ts|tsx|js|jsx)$/,
+        exclude: /node_modules/,
+        loader: require.resolve('babel-loader'),
+        options: {
+          presets: [
+            require.resolve('@babel/preset-env'),
+            require.resolve('@babel/preset-react'),
+            require.resolve('@babel/preset-typescript')
+          ]
+        }
+      }
     ]
   },
   performance: {
